+++ src/main/java/com/nhnacademy/inkbridge/backend/repository/custom/BookOrderDetailRepositoryCustom.java
@@ -3,7 +3,7 @@ package com.nhnacademy.inkbridge.backend.repository.custom;
 import com.nhnacademy.inkbridge.backend.dto.book.BookStockUpdateRequestDto;
 import com.nhnacademy.inkbridge.backend.dto.order.OrderDetailReadResponseDto;
 import com.nhnacademy.inkbridge.backend.entity.BookOrderDetail;
-import com.nhnacademy.inkbridge.backend.service.OrderBooksIdResponseDto;
+import com.nhnacademy.inkbridge.backend.dto.order.OrderBooksIdResponseDto;
 import java.util.List;
 
 /**
