+++ src/test/java/com/nhnacademy/inkbridge/backend/service/impl/BookOrderServiceImplTest.java
@@ -13,10 +13,7 @@ import static org.mockito.Mockito.verify;
 import com.nhnacademy.inkbridge.backend.dto.order.OrderCreateRequestDto.BookOrderCreateRequestDto;
 import com.nhnacademy.inkbridge.backend.dto.order.OrderCreateResponseDto;
 import com.nhnacademy.inkbridge.backend.dto.order.OrderPayInfoReadResponseDto;
-import com.nhnacademy.inkbridge.backend.dto.OrderedMemberPointReadResponseDto;
-import com.nhnacademy.inkbridge.backend.dto.order.OrderCreateRequestDto.BookOrderCreateRequestDto;
-import com.nhnacademy.inkbridge.backend.dto.order.OrderCreateResponseDto;
-import com.nhnacademy.inkbridge.backend.dto.order.OrderPayInfoReadResponseDto;
+import com.nhnacademy.inkbridge.backend.dto.order.OrderedMemberPointReadResponseDto;
 import com.nhnacademy.inkbridge.backend.dto.order.OrderReadResponseDto;
 import com.nhnacademy.inkbridge.backend.dto.order.OrderResponseDto;
 import com.nhnacademy.inkbridge.backend.entity.BookOrder;
